core: stdcm: split `VisitedNodes` into several files
No functional change there
