+++ core/src/main/kotlin/fr/sncf/osrd/stdcm/graph/visited_node_tracking/VisitedNodes.kt
@@ -1,15 +1,11 @@
-package fr.sncf.osrd.stdcm.graph
+package fr.sncf.osrd.stdcm.graph.VisitedNodeTracking
 
-import com.google.common.collect.Range
-import com.google.common.collect.RangeMap
-import com.google.common.collect.TreeRangeMap
-import com.google.common.collect.TreeRangeSet
 import fr.sncf.osrd.api.FullInfra
-import fr.sncf.osrd.envelope_sim.TrainPhysicsIntegrator.areTimesEqual
 import fr.sncf.osrd.sim_infra.api.BlockId
 import fr.sncf.osrd.sim_infra.api.DirDetectorId
 import fr.sncf.osrd.sim_infra.utils.getBlockEntry
 import fr.sncf.osrd.sim_infra.utils.getBlockExit
+import fr.sncf.osrd.stdcm.graph.TimeData
 import fr.sncf.osrd.stdcm.infra_exploration.EdgeIdentifier
 import fr.sncf.osrd.stdcm.infra_exploration.InfraExplorer
 import fr.sncf.osrd.stdcm.infra_exploration.getRemainingBlocks
@@ -72,95 +68,7 @@ data class VisitedNodes(
     }
 
     /** Any class that implements this interface may be added to the visited ranges. */
-    private sealed interface ConditionallyVisitedRange {
-        /** Is the given (sub)range covered by `this`? */
-        fun isVisited(
-            range: Range<Double>,
-            parameters: Parameters,
-        ): Boolean
-
-        /**
-         * Returns the range that is most likely to define elements as "already visited". Used to
-         * merge overlapping ranges in the range map.
-         */
-        fun mergeWith(other: ConditionallyVisitedRange?): ConditionallyVisitedRange
-    }
-
-    /**
-     * The given range is already visited by changing departure time. Such ranges are *always*
-     * already visited.
-     */
-    private data object VisitedWithDepartureTimeChange : ConditionallyVisitedRange {
-        override fun isVisited(
-            range: Range<Double>,
-            parameters: Parameters,
-        ): Boolean {
-            // If the other option were better, it would have been visited first
-            return true
-        }
-
-        override fun mergeWith(other: ConditionallyVisitedRange?): ConditionallyVisitedRange {
-            return this
-        }
-    }
-
-    /**
-     * The given range is already visited by adding stop durations. A candidate is already visited
-     * if their cost is higher or equal, or for equal cost if they add at least as much stop time as
-     * this.
-     */
-    private data class VisitedWithAddedStopTime(
-        // the time t is visited at f(t) seconds of total stop duration
-        private val visitedWithStopTime: LinearFunction,
-        // If the cost is higher, we don't even consider stop time
-        private val baseCost: Double,
-    ) : ConditionallyVisitedRange {
-        override fun isVisited(
-            range: Range<Double>,
-            parameters: Parameters,
-        ): Boolean {
-            if (parameters.nodeCost > baseCost) return true
-            val stopDurationForVisited = visitedWithStopTime.apply(range.lowerEndpoint())
-            return stopDurationForVisited <= parameters.timeData.totalStopDuration
-        }
-
-        override fun mergeWith(other: ConditionallyVisitedRange?): ConditionallyVisitedRange {
-            if (other is VisitedWithDepartureTimeChange) return other
-            if (other is VisitedWithAddedStopTime) {
-                if (!areTimesEqual(baseCost, other.baseCost))
-                    return if (baseCost < other.baseCost) this else other
-                return if (visitedWithStopTime < other.visitedWithStopTime) this else other
-            }
-            return this
-        }
-    }
-
-    /**
-     * The given range is already visited by adding margins. A candidate is already visited if their
-     * cost is higher or equal (accounting for any extra travel time).
-     */
-    private data class VisitedWithAddedTravelTime(
-        // the time t is visited at f(t) seconds of total running time duration
-        private val visitedWithTravelTime: LinearFunction,
-    ) : ConditionallyVisitedRange {
-        override fun isVisited(
-            range: Range<Double>,
-            parameters: Parameters,
-        ): Boolean {
-            val runningTimeForVisited = visitedWithTravelTime.apply(range.upperEndpoint())
-            return runningTimeForVisited <= parameters.nodeCost
-        }
-
-        override fun mergeWith(other: ConditionallyVisitedRange?): ConditionallyVisitedRange {
-            if (other is VisitedWithAddedTravelTime) {
-                return if (visitedWithTravelTime < other.visitedWithTravelTime) this else other
-            }
-            return other ?: this
-        }
-    }
-
-    private val visitedRangesPerLocation =
-        mutableMapOf<Fingerprint, RangeMap<Double, ConditionallyVisitedRange>>()
+    private val visitedRangesPerLocation = mutableMapOf<Fingerprint, VisitedRangeMap>()
 
     // For any given block, keep track of the maximum number of steps that have been reached.
     // Any future path that has less passed steps at this block will be discarded.
@@ -175,8 +83,7 @@ data class VisitedNodes(
     // First key is the number of visited steps.
     // Note: nodes are located at the *start* of blocks, so it's handled by accessing
     // the start of the "current" block.
-    private val visitedAtDetector =
-        mutableMapOf<Int, MutableMap<DirDetectorId, RangeMap<Double, ConditionallyVisitedRange>>>()
+    private val visitedAtDetector = mutableMapOf<Int, MutableMap<DirDetectorId, VisitedRangeMap>>()
 
     /** Returns true if the input has already been visited */
     fun isVisited(
@@ -221,14 +128,14 @@ data class VisitedNodes(
 
                 if (
                     mapAtDetector != null &&
-                        isMapVisited(mapAtDetector, parameters.copy(timeData = newTimeData))
+                        mapAtDetector.isVisited(parameters.copy(timeData = newTimeData))
                 ) {
                     return true
                 }
             }
         }
         val visitedRanges = visitedRangesPerLocation[parameters.fingerprint] ?: return false
-        return isMapVisited(visitedRanges, parameters)
+        return visitedRanges.isVisited(parameters)
     }
 
     /** Marks the input as visited */
@@ -241,7 +148,7 @@ data class VisitedNodes(
             minPassedStepsForBlock[parameters.explorer.getCurrentBlock()] =
                 fingerprint.waypointIndex
         }
-        val visitedRanges = visitedRangesPerLocation.getOrPut(fingerprint) { TreeRangeMap.create() }
+        val visitedRanges = visitedRangesPerLocation.getOrPut(fingerprint) { VisitedRangeMap() }
 
         val timeData = parameters.timeData
         val startTime = timeData.earliestReachableTime
@@ -259,8 +166,7 @@ data class VisitedNodes(
             startTime + timeData.maxDepartureDelayingWithoutConflict + minDelay
         val endRangeExtraTravelTime = endRangeExtraStopTime + parameters.maxMarginDuration
 
-        markAsVisited(
-            visitedRanges,
+        visitedRanges.markAsVisited(
             startTime,
             endRangeDepartureTimeChange,
             endRangeExtraStopTime,
@@ -283,10 +189,9 @@ data class VisitedNodes(
                         block,
                     )
                 ) {
-                    TreeRangeMap.create()
+                    VisitedRangeMap()
                 }
-            markAsVisited(
-                visitedRangesAtStartOfBlock,
+            visitedRangesAtStartOfBlock.markAsVisited(
                 startTime,
                 endRangeDepartureTimeChange,
                 endRangeExtraStopTime,
@@ -295,92 +200,5 @@ data class VisitedNodes(
                 parameters.nodeCost,
             )
         }
-    }
-
-    /**
-     * Marks the given time ranges as visited in the map. Adds 3 ranges: visited with added
-     * departure time, with extra stop duration, and with added travel time.
-     */
-    private fun markAsVisited(
-        map: RangeMap<Double, ConditionallyVisitedRange>,
-        startTime: Double,
-        endRangeDepartureTimeChange: Double,
-        endRangeExtraStopTime: Double,
-        endRangeExtraTravelTime: Double,
-        totalStopDuration: Double,
-        nodeCost: Double,
-    ) {
-        fun putRange(start: Double, end: Double, value: ConditionallyVisitedRange) {
-            if (start < end) {
-                val range = Range.closedOpen(start, end)
-                map.merge(range, value) { a, b -> a.mergeWith(b) }
-            }
-        }
-
-        // Visited with just departure time change, this is always considered as "visited".
-        // (The end of all ranges depends on conflicting occupancy along the path)
-        putRange(
-            startTime,
-            endRangeDepartureTimeChange,
-            VisitedWithDepartureTimeChange,
-        )
-        // Visited with extra stop duration, starting from the end of the previous range
-        putRange(
-            endRangeDepartureTimeChange,
-            endRangeExtraStopTime,
-            VisitedWithAddedStopTime(
-                LinearFunction(totalStopDuration - endRangeDepartureTimeChange),
-                nodeCost,
-            )
-        )
-        // Visited with extra margins, starting from the end of the previous range
-        putRange(
-            endRangeExtraStopTime,
-            endRangeExtraTravelTime,
-            VisitedWithAddedTravelTime(LinearFunction(nodeCost - endRangeExtraStopTime))
-        )
-    }
-
-    /**
-     * Returns true if the map already contains visited ranges for all new ranges from the given
-     * parameters.
-     */
-    private fun isMapVisited(
-        map: RangeMap<Double, ConditionallyVisitedRange>,
-        parameters: Parameters,
-    ): Boolean {
-        val timeData = parameters.timeData
-
-        val visitingRange =
-            Range.closedOpen(
-                timeData.earliestReachableTime,
-                timeData.earliestReachableTime + timeData.maxDepartureDelayingWithoutConflict
-            )
-        if (visitingRange.isEmpty) {
-            // Special case for empty range, `subRangeMap` returns an empty list
-            val value = map.get(timeData.earliestReachableTime) ?: return false
-            return value.isVisited(visitingRange, parameters)
-        }
-        val subMap = map.subRangeMap(visitingRange)
-
-        // Keep track of any range that isn't covered by the map
-        val uncovered = TreeRangeSet.create<Double>()
-        uncovered.add(visitingRange)
-        for (entry in subMap.asMapOfRanges()) {
-            uncovered.remove(entry.key)
-            // Value isn't visited: we can early return "false"
-            if (!entry.value.isVisited(entry.key, parameters)) return false
-        }
-        // If any area is left uncovered, we still return "false"
-        return uncovered.isEmpty
-    }
-}
-
-/** Linear function with slope of 1. Used to represent the cost to reach a given time with margin */
-private data class LinearFunction(private val y0: Double) : Comparable<LinearFunction> {
-    fun apply(x: Double): Double = y0 + x
-
-    override fun compareTo(other: LinearFunction): Int {
-        return y0.compareTo(other.y0)
     }
 }
