+++ core/src/test/kotlin/fr/sncf/osrd/stdcm/VisitedNodesTests.kt
@@ -2,7 +2,7 @@ package fr.sncf.osrd.stdcm
 
 import fr.sncf.osrd.stdcm.graph.StopTimeData
 import fr.sncf.osrd.stdcm.graph.TimeData
-import fr.sncf.osrd.stdcm.graph.VisitedNodes
+import fr.sncf.osrd.stdcm.graph.VisitedNodeTracking.VisitedNodes
 import fr.sncf.osrd.stdcm.infra_exploration.EdgeIdentifier
 import fr.sncf.osrd.utils.CachedBlockMRSPBuilder
 import fr.sncf.osrd.utils.DummyInfra
